backend/pyproject.toml
[tool.poetry]
name = "D4CAPI"
version = "0.1.0"
description = "Small api used for the sensors in D4C"
authors = ["Dan Amador"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.6.1"
gunicorn = "^19.9"
fastapi = "^0.13.0"
pydantic = {version = "^0.21.0",extras = ["email"]}
passlib = {version = "^1.7",extras = ["bcrypt"]}
pyjwt = "^1.7"
asyncpg = "^0.18.3"
alembic = "^1.0"
psycopg2-binary = "^2.8"
python-slugify = "^3.0"
Unidecode = "^1.0"
python-dotenv = "^0.10.1"
databases = "^0.2.1"
shapely = "^1.6"
sqlalchemy_utils = "^0.34.2"
geoalchemy2 = "^0.6.3"
python-multipart = "^0.0.5"
numpy = "^1.17"
requests = "^2.22"
librosa = "^0.7.0"


[tool.poetry.dev-dependencies]

[tool.black]
exclued = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

